cleaned and added test
