+++ QEfficient/utils/export_utils.py
@@ -6,6 +6,7 @@
 # -----------------------------------------------------------------------------
 
 import copy
+import contextlib
 import inspect
 import os
 import re
@@ -13,12 +14,131 @@ import warnings
 from pathlib import Path
 from typing import Dict
 
+import torch
+
 from QEfficient.base.onnx_transforms import CustomOpTransform, RenameFunctionOutputsTransform
 from QEfficient.transformers.cache_utils import InvalidIndexProvider
 from QEfficient.utils.cache import QEFF_HOME
 from QEfficient.utils.hash_utils import create_export_hash
 from QEfficient.utils.logging_utils import logger
 from QEfficient.utils.torch_patches import apply_torch_patches, undo_torch_patches
+
+
+_SAFE_ONNX_EXPORT_PASS_NAMES = (
+    "_jit_pass_constant_propagation",
+    "_jit_pass_dce",
+    "_jit_pass_cse",
+    "_jit_pass_canonicalize_graph_fuser_ops",
+    "_jit_pass_peephole",
+    "_jit_pass_fuse_addmm",
+    "_jit_pass_onnx_eval_peephole",
+    "_jit_pass_onnx_constant_fold",
+    "_jit_pass_dce_allow_deleting_nodes_with_side_effects",
+    "_jit_pass_canonicalize",
+    "_jit_pass_onnx_graph_shape_type_inference",
+    "_jit_pass_onnx_deduplicate_initializers",
+)
+
+
+def _qeff_noop(*args, **kwargs):
+    return None
+
+
+def _qeff_false_noop(*args, **kwargs):
+    return False
+
+
+def _qeff_first_arg(*args, **kwargs):
+    return args[0] if args else None
+
+
+def _qeff_second_arg_or_empty_dict(*args, **kwargs):
+    if len(args) >= 2:
+        return args[1]
+    return kwargs.get("params_dict", {})
+
+
+_SAFE_ONNX_EXPORT_PASS_REPLACEMENTS = {
+    "_jit_pass_constant_propagation": _qeff_noop,
+    "_jit_pass_dce": _qeff_noop,
+    "_jit_pass_cse": _qeff_false_noop,
+    "_jit_pass_canonicalize_graph_fuser_ops": _qeff_noop,
+    "_jit_pass_peephole": _qeff_noop,
+    "_jit_pass_fuse_addmm": _qeff_noop,
+    "_jit_pass_onnx_eval_peephole": _qeff_second_arg_or_empty_dict,
+    "_jit_pass_onnx_constant_fold": _qeff_second_arg_or_empty_dict,
+    "_jit_pass_dce_allow_deleting_nodes_with_side_effects": _qeff_noop,
+    "_jit_pass_canonicalize": _qeff_first_arg,
+    "_jit_pass_onnx_graph_shape_type_inference": _qeff_noop,
+    "_jit_pass_onnx_deduplicate_initializers": _qeff_second_arg_or_empty_dict,
+}
+
+
+def _safe_onnx_export_passes_from_env(default_disable_safe_passes: bool = False) -> tuple[str, ...]:
+    disabled_passes = os.environ.get("QEFF_ONNX_DISABLE_SAFE_EXPORT_PASSES")
+    if disabled_passes is None:
+        return _SAFE_ONNX_EXPORT_PASS_NAMES if default_disable_safe_passes else ()
+
+    normalized_value = disabled_passes.strip().lower()
+    if normalized_value in {"", "0", "false", "no", "off"}:
+        return ()
+    if normalized_value in {"1", "true", "yes", "on", "safe", "all"}:
+        return _SAFE_ONNX_EXPORT_PASS_NAMES
+
+    requested_passes = []
+    for raw_name in disabled_passes.split(","):
+        pass_name = raw_name.strip()
+        if not pass_name:
+            continue
+        if not pass_name.startswith("_jit_pass_"):
+            pass_name = f"_jit_pass_{pass_name}"
+        if pass_name not in _SAFE_ONNX_EXPORT_PASS_REPLACEMENTS:
+            raise ValueError(
+                "QEFF_ONNX_DISABLE_SAFE_EXPORT_PASSES can only include safe passes. "
+                f"Unknown or unsafe pass: {raw_name!r}. "
+                f"Allowed passes: {', '.join(_SAFE_ONNX_EXPORT_PASS_NAMES)}"
+            )
+        requested_passes.append(pass_name)
+    return tuple(dict.fromkeys(requested_passes))
+
+
+@contextlib.contextmanager
+def _disable_safe_onnx_export_passes_from_env(default_disable_safe_passes: bool = False):
+    pass_names = _safe_onnx_export_passes_from_env(default_disable_safe_passes=default_disable_safe_passes)
+    if not pass_names:
+        yield
+        return
+
+    originals = []
+    for pass_name in pass_names:
+        if not hasattr(torch._C, pass_name):
+            raise AttributeError(f"torch._C has no ONNX export pass named {pass_name}")
+        originals.append((pass_name, getattr(torch._C, pass_name)))
+        setattr(torch._C, pass_name, _SAFE_ONNX_EXPORT_PASS_REPLACEMENTS[pass_name])
+    logger.info("Disabled safe torch.onnx export passes: %s", ", ".join(pass_names))
+    try:
+        yield
+    finally:
+        for pass_name, original in reversed(originals):
+            setattr(torch._C, pass_name, original)
+
+
+def _apply_onnx_export_env_kwargs(export_kwargs: Dict) -> None:
+    do_constant_folding = os.environ.get("QEFF_ONNX_DO_CONSTANT_FOLDING")
+    if do_constant_folding is None or "do_constant_folding" in export_kwargs:
+        return
+
+    normalized_value = do_constant_folding.strip().lower()
+    if normalized_value in {"1", "true", "yes", "on"}:
+        export_kwargs["do_constant_folding"] = True
+    elif normalized_value in {"0", "false", "no", "off"}:
+        export_kwargs["do_constant_folding"] = False
+    else:
+        raise ValueError(
+            "QEFF_ONNX_DO_CONSTANT_FOLDING must be one of "
+            "1/0, true/false, yes/no, or on/off. "
+            f"Got: {do_constant_folding!r}"
+        )
 
 
 def export_wrapper(func):
@@ -114,7 +234,8 @@ def _generate_export_hash(qeff_model, args, kwargs, func):
     bound_args.apply_defaults()
     all_args = bound_args.arguments
     export_kwargs = dict(all_args.get("export_kwargs") or {})
-    if func.__name__ == "_export_layerwise":
+    is_layerwise_export = func.__name__ == "_export_layerwise"
+    if is_layerwise_export:
         export_kwargs["_qeff_layerwise_export"] = True
 
     do_constant_folding_env = os.environ.get("QEFF_ONNX_DO_CONSTANT_FOLDING")
@@ -124,6 +245,8 @@ def _generate_export_hash(qeff_model, args, kwargs, func):
     disabled_safe_passes_env = os.environ.get("QEFF_ONNX_DISABLE_SAFE_EXPORT_PASSES")
     if disabled_safe_passes_env is not None:
         export_kwargs["_qeff_env_onnx_disable_safe_export_passes"] = disabled_safe_passes_env.strip().lower()
+    elif is_layerwise_export:
+        export_kwargs["_qeff_layerwise_default_disable_safe_export_passes"] = True
 
     if export_kwargs:
         all_args["export_kwargs"] = export_kwargs
